+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Drop `epfl_si.rancher.rke2_access` role, as it was unused and at odds with 0.6.0's strategy of not managing kubeconfig files anymore
 - `epfl_si.rancher.rke2-node` role
 - `epfl_si.rancher.rancher_project` lookup plugin
+- `epfl_si.rancher.rancher_helm_chart` action plugin
 
 # Version 0.6.0: major bugfix and API change release